pyproject.toml
[project]
name = "informer-bot"
version = "0.1.0"
description = "Telegram news aggregator that summarises channel posts with Claude"
requires-python = ">=3.12"
dependencies = [
    "telethon>=1.36",
    "python-telegram-bot>=21.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "python-dotenv>=1.0",
    "aiohttp>=3.10",
]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "mypy>=1.20.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[[tool.mypy.overrides]]
module = ["telethon.*"]
ignore_missing_imports = true

[tool.uv]
package = false
